feat: eq + ord (and partials) for HiddenValue
Implementing raw over the struct rather than
for either just value or obf_string.
If someone wants to compare the subvalue they can
do that manually.  However, as default behavior,
one would expect equality to indicate equality of
the struct.  (And similar, less important, arg for
ord).  Having a default impl is important to allow
default impls on objects with a HiddenValue field.